PRSD-927: Searches LandlordWithListedPropertyCountRepository instead of LandlordRepository
